fix : instance injection

+++ src/user/user.controller.ts
@@ -2,15 +2,13 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Headers,
-  Inject,
   NotFoundException,
   Param,
   Post,
-  Req,
 } from '@nestjs/common';
-import { Request } from 'express';
 import { CookieService } from 'src/cookie/cookie.service';
 import { UserService } from './user.service';
 
@@ -20,76 +18,93 @@ export class UserController {
     private readonly userService: UserService,
     private readonly cookieService: CookieService,
   ) {}
+
+  // {
+  //   "id": 20,
+  //   "intraid": "alee",
+  //   "avatar": "https://i.imgflip.com/1rpfag.jpg",
+  //   "rating": 1205,
+  //   "wincount": 10,
+  //   "losecount": 5,
+  //   "email": "alee@gmail.com",
+  //   "isotp": true
+  // }
   @Get()
-  async getUser(
-    @Param('id') intraID: string,
-    @Headers('cookie') cookieRaw: string,
-  ) {
-    console.log(`cookieRaw : ${cookieRaw}`);
+  async getUser(@Headers('cookie') cookieRaw: string) {
     const cookie = this.cookieService.extractCookie(cookieRaw);
-    console.log(`client Key : ${this.userService.getIntraID(cookie)}`); //??
-    console.log(`cookie Key : ${cookie}`);
-    // console.log(cookie);//
-    // console.log(`who : ${this.userService.getIntraID(cookie)}`);
-    const userData = await this.userService.findUser(intraID);
-    if (userData == null) throw new NotFoundException(`${intraID} not found.`);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    const userData = await this.userService.findUser(target);
+    if (userData == null) throw new NotFoundException(`client not found.`);
     return userData;
   }
 
   // {
-  //   "intraID": "hena",
   //   "result": {
   //     "win": false,
   //     "lose": true
   //   }
   // }
   @Post('/result')
-  async updateResult(@Body('intraID') intraID: string, @Body('result') result) {
-    const updateResult = await this.userService.updateResult(intraID, result);
+  async updateResult(
+    @Headers('cookie') cookieRaw: string,
+    @Body('result') result,
+  ) {
+    const cookie = this.cookieService.extractCookie(cookieRaw);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    const updateResult = await this.userService.updateResult(target, result);
     if (updateResult == false)
-      throw new NotFoundException(`${intraID} not found.`);
+      throw new NotFoundException(`${target} not found.`);
     return;
   }
 
   // {
-  //   "intraID": "alee",
   //   "email": "alee@gmail.com"
   // }
-  @Post('/email')
+  @Post('/email') //ok
   async updateEmail(
-    @Body('intraID') intraID: string,
+    @Headers('cookie') cookieRaw: string,
     @Body('email') email: string,
   ) {
-    const updateEmail = await this.userService.updateEmail(intraID, email);
+    const cookie = this.cookieService.extractCookie(cookieRaw);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    const updateEmail = await this.userService.updateEmail(target, email);
     if (updateEmail == false)
-      throw new NotFoundException(`${intraID} not found.`);
+      throw new NotFoundException(`${target} not found.`);
     return;
   }
 
   // {
-  //   "intraID": "hena",
-  //   "opt": true
+  //   "otp": true
   // }
-  @Post('/opt')
-  async updateOpt(@Body('intraID') intraID: string, @Body('opt') opt: boolean) {
-    const updateOpt = await this.userService.updateOpt(intraID, opt);
-    if (updateOpt == false)
-      throw new NotFoundException(`${intraID} not found.`);
+  @Post('/otp')
+  async updateOtp(
+    @Headers('cookie') cookieRaw: string,
+    @Body('otp') otp: boolean,
+  ) {
+    const cookie = this.cookieService.extractCookie(cookieRaw);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    const updateOtp = await this.userService.updateOtp(target, otp);
+    if (updateOtp == false) throw new NotFoundException(`${target} not found.`);
     return;
   }
 
   // {
-  //   "intraID": "hena",
   //   "url": "https://i.imgflip.com/1rpfag.jpg"
   // }
   @Post('/avatar')
   async updateAvatarURL(
-    @Body('intraID') intraID: string,
+    @Headers('cookie') cookieRaw: string,
     @Body('url') url: string,
   ) {
-    const updateURL = await this.userService.updateURL(intraID, url);
-    if (updateURL == false)
-      throw new NotFoundException(`${intraID} not found.`);
+    const cookie = this.cookieService.extractCookie(cookieRaw);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    const updateURL = await this.userService.updateURL(target, url);
+    if (updateURL == false) throw new NotFoundException(`${target} not found.`);
     return;
   }
 
@@ -97,7 +112,15 @@ export class UserController {
   //   "intraID": "hena",
   // }
   @Delete('/:id')
-  async deleteUser(@Param('id') intraID: string) {
+  async deleteUser(
+    @Headers('cookie') cookieRaw: string,
+    @Param('id') intraID: string,
+  ) {
+    console.log(intraID);
+    const cookie = this.cookieService.extractCookie(cookieRaw);
+    if (cookie == undefined) throw new NotFoundException('cookie not found');
+    const target = this.userService.getIntraID(cookie);
+    if (intraID != target) throw new ForbiddenException('forbbiden request.');
     const delResult = await this.userService.deleteUser(intraID);
     if (delResult == false)
       throw new NotFoundException(`${intraID} not found.`);
